+++ apm-sniffer/apm-sdk-plugin/mongodb-3.x-plugin/src/main/java/org/apache/skywalking/apm/plugin/mongodb/v3/define/v37/MongoDBOperationExecutorInstrumentation.java
@@ -28,7 +28,7 @@ import org.apache.skywalking.apm.agent.core.plugin.interceptor.InstanceMethodsIn
 import org.apache.skywalking.apm.agent.core.plugin.interceptor.enhance.ClassInstanceMethodsEnhancePluginDefine;
 import org.apache.skywalking.apm.agent.core.plugin.match.ClassMatch;
 import org.apache.skywalking.apm.agent.core.plugin.match.NameMatch;
-import org.apache.skywalking.apm.plugin.mongodb.v3.interceptor.v37.MongoDBV37OperationExecutorInterceptor;
+import org.apache.skywalking.apm.plugin.mongodb.v3.interceptor.v37.MongoDBOperationExecutorInterceptor;
 
 /**
  * {@code com.mongodb.client.internal.OperationExecutor} which is unified entrance of execute mongo command.
@@ -38,16 +38,16 @@ import org.apache.skywalking.apm.plugin.mongodb.v3.interceptor.v37.MongoDBV37Ope
  * support: 3.7.x
  *
  * @author scolia
- * @see MongoDBV37OperationExecutorInterceptor
+ * @see MongoDBOperationExecutorInterceptor
  */
 @SuppressWarnings({"Duplicates"})
-public class MongoDBV37OperationExecutorInstrumentation extends ClassInstanceMethodsEnhancePluginDefine {
+public class MongoDBOperationExecutorInstrumentation extends ClassInstanceMethodsEnhancePluginDefine {
 
     private static final String WITNESS_CLASS = "com.mongodb.client.internal.MongoClientDelegate";
 
     private static final String ENHANCE_CLASS = "com.mongodb.client.internal.MongoClientDelegate$DelegateOperationExecutor";
 
-    private static final String INTERCEPTOR_CLASS = "org.apache.skywalking.apm.plugin.mongodb.v3.interceptor.v37.MongoDBV37OperationExecutorInterceptor";
+    private static final String INTERCEPTOR_CLASS = "org.apache.skywalking.apm.plugin.mongodb.v3.interceptor.v37.MongoDBOperationExecutorInterceptor";
 
     private static final String METHOD_NAME = "execute";
