+++ apm-sniffer/apm-sdk-plugin/mongodb-3.x-plugin/src/main/java/org/apache/skywalking/apm/plugin/mongodb/v3/interceptor/v30/MongoDBInterceptor.java
@@ -17,33 +17,48 @@
  */
 
 
-package org.apache.skywalking.apm.plugin.mongodb.v3.interceptor.v37;
+package org.apache.skywalking.apm.plugin.mongodb.v3.interceptor.v30;
 
+import com.mongodb.connection.Cluster;
 import org.apache.skywalking.apm.agent.core.context.ContextManager;
 import org.apache.skywalking.apm.agent.core.context.trace.AbstractSpan;
 import org.apache.skywalking.apm.agent.core.logging.api.ILog;
 import org.apache.skywalking.apm.agent.core.logging.api.LogManager;
 import org.apache.skywalking.apm.agent.core.plugin.interceptor.enhance.EnhancedInstance;
+import org.apache.skywalking.apm.agent.core.plugin.interceptor.enhance.InstanceConstructorInterceptor;
 import org.apache.skywalking.apm.agent.core.plugin.interceptor.enhance.InstanceMethodsAroundInterceptor;
 import org.apache.skywalking.apm.agent.core.plugin.interceptor.enhance.MethodInterceptResult;
+import org.apache.skywalking.apm.plugin.mongodb.v3.support.MongoRemotePeerHelper;
 import org.apache.skywalking.apm.plugin.mongodb.v3.support.MongoSpanHelper;
 
 import java.lang.reflect.Method;
 
 /**
+ * Intercept method of {@code com.mongodb.Mongo#execute(ReadOperation, ReadPreference)} or
+ * {@code com.mongodb.Mongo#execute(WriteOperation)}. record the MongoDB host, operation name and the key of the
+ * operation.
+ * <p>
+ * only supported: 3.0.x-3.5.x
+ *
  * @author scolia
  */
-@SuppressWarnings("Duplicates")
-public class MongoDBV37OperationExecutorInterceptor implements InstanceMethodsAroundInterceptor {
+@SuppressWarnings({"deprecation", "Duplicates"})
+public class MongoDBInterceptor implements InstanceMethodsAroundInterceptor, InstanceConstructorInterceptor {
 
-    private static final ILog logger = LogManager.getLogger(MongoDBV37OperationExecutorInterceptor.class);
+    private static final ILog logger = LogManager.getLogger(MongoDBInterceptor.class);
 
     @Override
-    public void beforeMethod(EnhancedInstance objInst, Method method, Object[] allArguments, Class<?>[] argumentsTypes,
-                             MethodInterceptResult result) {
+    public void onConstruct(EnhancedInstance objInst, Object[] allArguments) {
+        Cluster cluster = (Cluster) allArguments[0];
+        String peers = MongoRemotePeerHelper.getRemotePeer(cluster);
+        objInst.setSkyWalkingDynamicField(peers);
+    }
+
+
+    @Override
+    public void beforeMethod(EnhancedInstance objInst, Method method, Object[] allArguments,
+                             Class<?>[] argumentsTypes, MethodInterceptResult result) {
         String executeMethod = allArguments[0].getClass().getSimpleName();
-        // OperationExecutor has be mark it's remotePeer
-        // @see: MongoDBV37ClientDelegateInterceptor.afterMethod
         String remotePeer = (String) objInst.getSkyWalkingDynamicField();
         if (logger.isDebugEnable()) {
             logger.debug("Mongo execute: [executeMethod: {}, remotePeer: {}]", executeMethod, remotePeer);
@@ -52,8 +67,8 @@ public class MongoDBV37OperationExecutorInterceptor implements InstanceMethodsAr
     }
 
     @Override
-    public Object afterMethod(EnhancedInstance objInst, Method method, Object[] allArguments, Class<?>[] argumentsTypes,
-                              Object ret) {
+    public Object afterMethod(EnhancedInstance objInst, Method method, Object[] allArguments,
+                              Class<?>[] argumentsTypes, Object ret) {
         ContextManager.stopSpan();
         return ret;
     }
@@ -65,6 +80,4 @@ public class MongoDBV37OperationExecutorInterceptor implements InstanceMethodsAr
         activeSpan.errorOccurred();
         activeSpan.log(t);
     }
-
-
 }
